Intel-SA-00086-Recovery-Tool: Fix for downloads of cached expired CRL
Disable server-side cache to ensure latest certificates and CRL are
downloaded. In this case, Wget will send the remote server an appropriate
directive (Pragma: no-cache) to get the file from the remote service,
rather than returning the cached version.  This is especially useful
for retrieving and flushing out-of-date documents on proxy servers.

Signed-off-by: Imran Desai <imran.desai@intel.com>